every table has change to virtualized in snapp
