+++ apps/sensenet/src/components/content-list/actions-field.tsx
@@ -4,23 +4,18 @@ import MoreHoriz from '@material-ui/icons/MoreHoriz'
 
 export const ActionsField: React.FC<{
   onOpen: (ev: React.MouseEvent<HTMLButtonElement>) => void
-  virtual?: boolean
-}> = ({ onOpen, virtual }) => {
+}> = ({ onOpen }) => {
   return (
     <TableCell
       component="div"
-      style={
-        virtual
-          ? {
-              height: '57px',
-              width: '100%',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              padding: 0,
-            }
-          : { width: '64px' }
-      }>
+      style={{
+        height: '57px',
+        width: '100%',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+        padding: 0,
+      }}>
       <IconButton onClick={onOpen}>
         <MoreHoriz />
       </IconButton>
