+++ apps/sensenet/src/components/content-list/description-field.tsx
@@ -1,14 +1,17 @@
 import { TableCell } from '@material-ui/core'
 import React from 'react'
 
-export const DescriptionField: React.FC<{ text: string; virtual?: boolean }> = ({ text, virtual }) => (
+export const DescriptionField: React.FC<{ text: string }> = ({ text }) => (
   <TableCell
     component="div"
-    style={
-      virtual
-        ? { height: '57px', display: 'flex', justifyContent: 'center', alignItems: 'center', width: '100%', padding: 0 }
-        : {}
-    }>
+    style={{
+      height: '57px',
+      display: 'flex',
+      justifyContent: 'left',
+      alignItems: 'center',
+      width: '100%',
+      padding: 0,
+    }}>
     <div style={{ maxWidth: '300px' }}>{text ? text.replace(/<(.|\n)*?>/g, '') : ''}</div>
   </TableCell>
 )
