+++ apps/sensenet/src/components/content-list/display-name-field.tsx
@@ -12,12 +12,7 @@ type DisplayNameProps = {
   virtual?: boolean
 }
 
-export const DisplayNameComponent: React.FunctionComponent<DisplayNameProps> = ({
-  content,
-  device,
-  isActive,
-  virtual,
-}) => {
+export const DisplayNameComponent: React.FunctionComponent<DisplayNameProps> = ({ content, device, isActive }) => {
   const buttonRef = useRef<HTMLButtonElement>(null)
   const [isOpened, setIsOpened] = useState(false)
 
@@ -31,7 +26,7 @@ export const DisplayNameComponent: React.FunctionComponent<DisplayNameProps> = (
     <TableCell
       component="div"
       padding={'none'}
-      style={virtual ? { height: '57px', display: 'flex', alignItems: 'center', width: '100%', padding: 0 } : {}}>
+      style={{ height: '57px', display: 'flex', alignItems: 'center', width: '100%', padding: 0 }}>
       <div
         style={{
           display: 'flex',
