+++ apps/sensenet/src/components/content-list/reference-field.tsx
@@ -3,23 +3,19 @@ import React from 'react'
 import { GenericContent } from '@sensenet/default-content-types'
 import { Icon } from '../Icon'
 
-export const ReferenceField: React.FC<{ content: GenericContent; virtual?: boolean }> = ({ content, virtual }) => {
+export const ReferenceField: React.FC<{ content: GenericContent }> = ({ content }) => {
   return (
     <TableCell
       component="div"
       padding={'none'}
-      style={
-        virtual
-          ? {
-              height: '57px',
-              width: '100%',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              padding: 0,
-            }
-          : {}
-      }>
+      style={{
+        height: '57px',
+        width: '100%',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+        padding: 0,
+      }}>
       {content.Name !== 'Somebody' ? (
         <div style={{ display: 'flex', alignItems: 'center' }}>
           <Icon item={content} />
