+++ apps/sensenet/src/components/dashboard/query-widget.tsx
@@ -14,8 +14,9 @@ import {
 } from '@sensenet/hooks-react'
 import { QueryWidget as QueryWidgetModel } from '../../services/PersonalSettings'
 import { useContentRouting, useLocalization, useSelectionService, useStringReplace } from '../../hooks'
-import { CollectionComponent, isReferenceField } from '../content-list'
+import { isReferenceField } from '../content-list'
 import { encodeQueryData } from '../search'
+import { ReactVirtualizedTable } from '../content-list/react-virtualized-table'
 
 const QueryWidget: React.FunctionComponent<QueryWidgetModel<GenericContent> & RouteComponentProps> = props => {
   const [items, setItems] = useState<GenericContent[]>([])
@@ -133,7 +134,7 @@ const QueryWidget: React.FunctionComponent<QueryWidgetModel<GenericContent> & Ro
                   setLoadSettings: () => ({}),
                   setLoadAncestorsSettings: () => ({}),
                 }}>
-                <CollectionComponent
+                <ReactVirtualizedTable
                   disableSelection={!props.settings.enableSelection}
                   hideHeader={!props.settings.showColumnNames}
                   fieldsToDisplay={props.settings.columns}
